requirements.txt
gunicorn
Flask
pymongo
dnspython
python-dotenv
requests==2.25.1
python-bsonjs
# uwsgi

